add card gird system and add some package

+++ Client/Assets/Functions/World/Board/Scripts/BoardFaction.cs
@@ -10,9 +10,10 @@ namespace World.Board
         [System.Serializable]
         public struct Position
         {
-            public byte Index;
+            public int Index;  // Đổi từ byte sang int
             public GameObject ObjCard;
         }
+
 
         [SerializeField] private TextMeshProUGUI _txRemainingTimeSecond; //turn will have time limit
         [SerializeField] private Image _imageFrame; //skin just apply for owner faction
